Agentic Business Profit Intelligence tool/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
prefect>=2.14.0
openai>=1.0.0
fpdf2
ydata_profiling>=4.17.0